Test stream expectation function panic cases
Due to relaxed type constraints and the use of the `reflect` package to
determine the type of some values, there are scenarios that may result
in panics.

In a future version of Go (possibly 1.17) these functions may be updated
using generics to add further constraints and remove the need for
reflection.